+++ changelog.md
@@ -24,6 +24,7 @@
     * leash position isn't correct
   * 0.4.4-alpha:
   * Added dynamic crosshair function
+  * Added clip to space function
 
 * 0.3.0-alpha:
 * Made the hit result of the player's crosshair match the actual position of the crosshair
@@ -83,6 +84,7 @@
     * 缰绳位置不正确
   * 0.4.4-alpha:
   * 添加了动态十字准心功能
+  * 添加了限制摄像头范围(Clip to space)的功能
 
 * 0.3.0-alpha:
 * 使玩家十字准心的命中结果与十字准心的实际位置相匹配